Converted LibraryViewCallback.java to Kotlin
