+++ app/src/main/java/org/kiwix/kiwixmobile/zimManager/fileselectView/FileSelectListState.kt
@@ -15,16 +15,21 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
  */
+package org.kiwix.kiwixmobile.zimManager.fileselectView
 
-package org.kiwix.kiwixmobile
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.SelectionMode
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.SelectionMode.NORMAL
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem
+import org.kiwix.kiwixmobile.core.zim_manager.fileselect_view.adapter.BooksOnDiskListItem.BookOnDisk
 
-import android.net.ConnectivityManager
-import org.kiwix.kiwixmobile.zimManager.NetworkState
-import org.kiwix.kiwixmobile.zimManager.NetworkState.CONNECTED
-import org.kiwix.kiwixmobile.zimManager.NetworkState.NOT_CONNECTED
-
-val ConnectivityManager.networkState: NetworkState
-  get() = if (activeNetworkInfo?.isConnected == true)
-    CONNECTED
-  else
-    NOT_CONNECTED
+data class FileSelectListState(
+  val bookOnDiskListItems: List<BooksOnDiskListItem>,
+  val selectionMode: SelectionMode = NORMAL
+) {
+  val selectedBooks by lazy {
+    bookOnDiskListItems.filter(
+      BooksOnDiskListItem::isSelected
+    )
+      .filterIsInstance(BookOnDisk::class.java)
+  }
+}
